+++ ap-java-sdk/src/main/java/com/alternativepayments/models/transaction/Transaction.java
@@ -1,7 +1,5 @@
 package com.alternativepayments.models.transaction;
 
-import java.math.BigDecimal;
-
 import org.joda.time.DateTime;
 
 import com.alternativepayments.models.BaseModel;
@@ -23,7 +21,7 @@ public class Transaction extends BaseModel {
     private final String customerId;
     private final Payment payment;
     private final String token;
-    private final BigDecimal amount;
+    private final int amount;
     private final String currency;
     private final String merchantPassThruData;
     private final String merchantTransactionId;
@@ -68,7 +66,7 @@ public class Transaction extends BaseModel {
             @JsonProperty("customerId") final String customerId,
             @JsonProperty("payment") final Payment payment,
             @JsonProperty("token") final String token,
-            @JsonProperty("amount") final BigDecimal amount,
+            @JsonProperty("amount") final int amount,
             @JsonProperty("currency") final String currency,
             @JsonProperty("merchantPassThruData") final String merchantPassThruData,
             @JsonProperty("merchantTransactionId") final String merchantTransactionId,
@@ -105,7 +103,7 @@ public class Transaction extends BaseModel {
         // Required parameters
         private final Payment payment;
         private final String token;
-        private final BigDecimal amount;
+        private final int amount;
         private final String currency;
 
         // Optional parameters - initialize with default values
@@ -129,7 +127,7 @@ public class Transaction extends BaseModel {
          * @param amount amount for the transaction
          * @param currency currency for the transaction
          */
-        public Builder(final Payment payment, final String token, final BigDecimal amount, final String currency) {
+        public Builder(final Payment payment, final String token, final int amount, final String currency) {
             this.payment = payment;
             this.token = token;
             this.amount = amount;
@@ -332,7 +330,7 @@ public class Transaction extends BaseModel {
     /**
      * @return amount for transaction.
      */
-    public BigDecimal getAmount() {
+    public int getAmount() {
         return amount;
     }
